Move optimization pass types into optimizer module
